pyproject.toml
[project]
name = "sifact"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "streamlit>=1.40.0",
    "uvicorn[standard]>=0.32.0",
    "beautifulsoup4>=4.12.0",
    "fastmcp>=3.2.3",
    "feedparser>=6.0.12",
    "gnews>=0.3.8",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
    "langgraph>=0.2.0",
    "lxml>=5.0.0",
    "newsapi-python>=0.2.7",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
    "tenacity>=9.0.0",
]
